Moving a bunch of class declarations back into header files. This is needed for later template instantiation.
